src/requirements.txt
discord.py==2.3.2
yt_dlp==2023.12.30
